refactor: extract HabitModal component
